requirements.txt
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
soupsieve==2.4.1
